# Spanify AntiFormat function in AvTrace, decrease allocations, improve perf