--- a/src/Microsoft.DotNet.Wpf/src/WindowsBase/MS/Internal/AvTrace.cs
+++ b/src/Microsoft.DotNet.Wpf/src/WindowsBase/MS/Internal/AvTrace.cs
@@ -375,37 +375,43 @@ namespace MS.Internal
 
 
         // replace { and } by {{ and }} - call if literal string will be passed to Format
-        static public string AntiFormat(string s)
+        public static string AntiFormat(string value)
         {
-            int formatIndex = s.IndexOfAny(FormatChars);
-            if (formatIndex < 0)
-                return s;
+            ReadOnlySpan<char> input = value.AsSpan();
 
-            StringBuilder sb = new StringBuilder();
-            int index = 0;
-            int lengthMinus1 = s.Length - 1;
+            // Check if there are any chars present
+            int formatIndex = input.IndexOfAny(FormatChars);
+            if (formatIndex == -1)
+                return value;
 
-            while (formatIndex >= 0)
+            StringBuilder sb = new(value.Length * 2);
+
+            while (input.Length > 0)
             {
-                if (formatIndex < lengthMinus1 && s[formatIndex] == s[formatIndex+1])
+                if (formatIndex == -1)
                 {
-                    // formatting character is already duplicated - leave them alone
-                    formatIndex = s.IndexOfAny(FormatChars, formatIndex+2);
+                    // No formatting character found, append rest of the string and exit
+                    sb.Append(input);
+                    break;
+                }
+                else if (input.Length > formatIndex + 1 && input[formatIndex] == input[formatIndex + 1])
+                {
+                    // Formatting character is already duplicated (append string only)
+                    sb.Append(input.Slice(0, formatIndex + 2));
+
+                    input = input.Slice(formatIndex + 2);
                 }
                 else
                 {
-                    // duplicate the formatting character
-                    sb.Append(s.AsSpan(index, formatIndex - index + 1));
-                    sb.Append(s[formatIndex]);
+                    // Duplicate the formatting character and append string
+                    sb.Append(input.Slice(0, formatIndex + 1));
+                    sb.Append(input[formatIndex]);
 
-                    index = formatIndex + 1;
-                    formatIndex = s.IndexOfAny(FormatChars, index);
+                    input = input.Slice(formatIndex + 1);
                 }
-            }
 
-            if (index <= lengthMinus1)
-            {
-                sb.Append(s.AsSpan(index));
+                // Find the next formatting character
+                formatIndex = input.IndexOfAny(FormatChars);
             }
 
             return sb.ToString();
@@ -491,14 +497,15 @@ namespace MS.Internal
         // Cache used by IsWpfTracingEnabledInRegistry
         private static Nullable<bool> _enabledInRegistry = null;
 
-        private static char[] FormatChars = new char[]{ '{', '}' };
+        private static ReadOnlySpan<char> FormatChars => ['{', '}'];
+
     }
 
     internal delegate void AvTraceEventHandler( AvTraceBuilder traceBuilder, object[] parameters, int start );
 
     internal class AvTraceBuilder
     {
-        private StringBuilder  _sb;
+        private readonly StringBuilder _sb;
 
         public AvTraceBuilder()
         {
@@ -527,25 +534,25 @@ namespace MS.Internal
 
         public void AppendFormat( string message, object arg1 )
         {
-            _sb.AppendFormat( CultureInfo.InvariantCulture, message, new object[] { AvTrace.ToStringHelper(arg1) } );
+            _sb.AppendFormat(CultureInfo.InvariantCulture, message, AvTrace.ToStringHelper(arg1));
         }
 
         public void AppendFormat( string message, object arg1, object arg2 )
         {
-            _sb.AppendFormat( CultureInfo.InvariantCulture, message, new object[] { AvTrace.ToStringHelper(arg1), AvTrace.ToStringHelper(arg2) } );
+            _sb.AppendFormat(CultureInfo.InvariantCulture, message, AvTrace.ToStringHelper(arg1), AvTrace.ToStringHelper(arg2));
         }
 
         public void AppendFormat( string message, string arg1 )
         {
-            _sb.AppendFormat( CultureInfo.InvariantCulture, message, new object[] { AvTrace.AntiFormat(arg1) } );
+            _sb.AppendFormat(CultureInfo.InvariantCulture, message, AvTrace.AntiFormat(arg1));
         }
 
         public void AppendFormat( string message, string arg1, string arg2 )
         {
-            _sb.AppendFormat( CultureInfo.InvariantCulture, message, new object[] { AvTrace.AntiFormat(arg1), AvTrace.AntiFormat(arg2) } );
+            _sb.AppendFormat(CultureInfo.InvariantCulture, message, AvTrace.AntiFormat(arg1), AvTrace.AntiFormat(arg2));
         }
 
-        public override string ToString( )
+        public override string ToString()
         {
             return _sb.ToString();
         }