## Description

Improves performance of the `AntiFormat` function in `AvTrace`, reducing allocations and improving the overall performance up to 2 times in some cases. Also changes `FormatChars` from `char[]` to `ReadOnlySpan<char>`, occurrence I've missed in #9230.

Second commit features added documentation for the function. I've included some basic asserts at the **Testing** section.

### Small strings within `StringBuilder` default capacity

| Method     | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Allocated [B] |
|----------- |----------:|-----------:|------------:|-------:|--------------:|
| PR__Edit |  236.1 ns |    4.75 ns |     5.84 ns | 0.0730 |        1224 B |
| Original |  400.7 ns |    8.04 ns |    15.29 ns | 0.1040 |        1744 B |

<details><summary>Benchmark code</summary>

```c#
AntiFormat2("{aaaa}");
AntiFormat2("{aaaa}}");
AntiFormat2("{{aaaa}}");
AntiFormat2("{aa{{aa}}");
AntiFormat2("{aa{{aa}}ab}c}d");
AntiFormat2("aaaa}");
AntiFormat2("aaaa");
AntiFormat2("a{{{aaa");
AntiFormat2("a{{{aaa}}}");

```

</details>

### Large strings, mostly without any replacements

| Method     | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Allocated [B] |
|----------- |----------:|-----------:|------------:|-------:|--------------:|
| PR__Edit|  127.5 ns |    1.29 ns |     1.08 ns | 0.0880 |        1472 B |
| Original |  242.9 ns |    4.87 ns |     5.00 ns | 0.1094 |        1832 B |

<details><summary>Benchmark code</summary>

```c#
AntiFormat("aaaaxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
AntiFormat("aaaaxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
AntiFormat("aaaxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxa");
AntiFormat("aaaaxxxc ccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc");
AntiFormat("aaaaabcd}yyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyy");
AntiFormat("aaaa}xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
AntiFormat("aaaa}6666666666666666666666666666666666666666666666666666666666666");
AntiFormat("aaaa7777777777777777777777777777777777777777777777777777777777777");
AntiFormat("aaaaggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggg");
```

</details>

### Large strings, with replacements (up to two times the perf)

| Method     | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Allocated [B] |
|----------- |----------:|-----------:|------------:|-------:|--------------:|
| PR__Edit|  394.7 ns |    7.73 ns |    10.05 ns | 0.3271 |        5472 B |
| Original|  798.9 ns |   11.30 ns |    10.57 ns | 0.4425 |        7408 B |

<details><summary>Benchmark code</summary>

```c#
AntiFormat("aaaaxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}");
AntiFormat("aaaaxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}");
AntiFormat("aaaxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxa}");
AntiFormat("aaaaxxxc ccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc}");
AntiFormat("aaaaabcd}yyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyy");
AntiFormat("aaaa}xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
AntiFormat("aaaa}6666666666666666666666666666666666666666666666666666666666666");
AntiFormat("aaaa7777777777777777777777777777777777777777777777777777777777777{");
AntiFormat("aaaaggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggg}");
AntiFormat("aaaa77777777777777777777777777{}777777777777777777777777777777777{");
AntiFormat("aaaaggggggggggggggggggggggggggggggggggggggggggggggggg{gggggggggggggg}");
```

</details>

## Customer Impact

Improved performance, decreased allocations.

## Regression

No.

## Testing

Local build, base testing between the two functions.

```c#
Assert.AreEqual(AntiFormat("{aaaa}}"), AntiFormat2("{aaaa}}"));
Assert.AreEqual(AntiFormat("{aa{{aa}}"), AntiFormat2("{aa{{aa}}"));
Assert.AreEqual(AntiFormat("{aa{{aa}}ab}c}d"), AntiFormat2("{aa{{aa}}ab}c}d"));
Assert.AreEqual(AntiFormat("aaaa}"), AntiFormat2("aaaa}"));
Assert.AreEqual(AntiFormat("aaaa"), AntiFormat2("aaaa"));
Assert.AreEqual(AntiFormat("a{{{aaa"), AntiFormat2("a{{{aaa"));
Assert.AreEqual(AntiFormat("a{{{aaa}}}"), AntiFormat2("a{{{aaa}}}"));
Assert.AreEqual(AntiFormat("{aaaa}"), AntiFormat2("{aaaa}"));
Assert.AreEqual(AntiFormat("{{aaaa}}"), AntiFormat2("{{aaaa}}"));
Assert.AreEqual(AntiFormat("{{{}aaaa}}"), AntiFormat2("{{{}aaaa}}"));
Assert.AreEqual(AntiFormat(string.Empty), AntiFormat2(string.Empty));
Assert.AreEqual(AntiFormat("a{{{}aaaa}}"), AntiFormat2("a{{{}aaaa}}"));
Assert.AreEqual(AntiFormat("a}aaaa}"), AntiFormat2("a}aaaa}"));
Assert.AreEqual(AntiFormat("}aaaa}"), AntiFormat2("}aaaa}"));
```

## Risk

Low.

 ###### Microsoft Reviewers: [Open in CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/dotnet/wpf/pull/9361)